Added final changes to etl.py
I added the final changes to the etl.py script so that it looks like the
EMR spark notebook in which I ran all of the commands.
Then, I added instructions in the README file, as well as a decription
of the different files in the repo, and some problems I ran into.